+++ src/Sync/Processor/IObjectSyncProcessor.php
@@ -2,6 +2,8 @@
 
 namespace srag\Plugins\Hub2\Sync\Processor;
 
+use srag\Plugins\Hub2\Exception\HubException;
+use srag\Plugins\Hub2\Exception\ILIASObjectNotFoundException;
 use srag\Plugins\Hub2\Object\DTO\IDataTransferObject;
 use srag\Plugins\Hub2\Object\IObject;
 use srag\Plugins\Hub2\Sync\IDataTransferObjectSort;
@@ -28,6 +30,9 @@ interface IObjectSyncProcessor {
 	 * @param IObject             $object
 	 * @param IDataTransferObject $dto
 	 * @param bool                $force Update all Objects without Hash comparison
+	 *
+	 * @throws ILIASObjectNotFoundException
+	 * @throws HubException
 	 */
 	public function process(IObject $object, IDataTransferObject $dto, bool $force = false);