+++ tests/Sync/Processor/GroupSyncProcessorTest.php
@@ -122,7 +122,7 @@ class GroupSyncProcessorTest extends AbstractSyncProcessorTests {
 	public function test_update_group_with_default_properties() {
 		$processor = new GroupSyncProcessor($this->origin, $this->originImplementation, $this->statusTransition, $this->originLog, $this->originNotifications, $this->activities);
 
-		$this->iobject->shouldReceive('updateStatus')->once()->with(IObject::STATUS_NOTHING_TO_UPDATE);
+		//$this->iobject->shouldReceive('updateStatus')->once()->with(IObject::STATUS_NOTHING_TO_UPDATE);
 
 		$this->iobject->shouldReceive('getStatus')->andReturn(IObject::STATUS_TO_UPDATE);
 		$this->iobject->shouldReceive('setData')->once()->with($this->dto->getData());