+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GlobalHook feature
 - New logs will be logged in database, not in files anymore
 - More meaningful GUI for force update synch: Buttons. Also implement this for a single origin synch
+- New origin synch status model
 - Fix: Add correct namespace for new origins
 - Fix: Not delete data on origin delete
 - Fix: Recreate deleted users